requirements.txt
    aiohttp==3.13.2
    aiosignal==1.4.0
    asgiref==3.11.0
    attrs==25.4.0
    certifi==2025.11.12
    charset-normalizer==3.4.4
    Django==5.2.8
    et-xmlfile==2.0.0
    frozenlist==1.8.0
    geoip2==5.2.0
    idna==3.11
    maxminddb==3.0.0
    multidict==6.7.0
    numpy==2.3.5
    openpyxl==3.1.5
    pandas==2.3.3
    polib==1.2.0
    python-dateutil==2.9.0.post0
    pytz==2025.2
    requests==2.32.5
    six==1.17.0
    sqlparse==0.5.3
    tendo==0.3.0
    tzdata==2025.2
    urllib3==2.5.0
    XlsxWriter==3.2.9
    yarl==1.22.0
    psycopg2-binary
    django-unfold
    django-environ
    gunicorn
    redis
    django-redis

